Throw error if GPT2 checkpoint fails to load
